cast insterted and included stdlib
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@12405 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ guienv/private/winnt/external/rotatetext.cpp
@@ -1,6 +1,7 @@
 #include <windows.h>
 #include <Winuser.h>
 #include <stdio.h>
+#include <stdlib.h>
 #include <math.h>                    // for trigonometry functions
 
 extern "C" 
@@ -119,23 +120,23 @@ extern "C"
 	      }
 	    else if ((0 < angle) && (angle < 90))
 	      {
-		x = x + abs(deltaX);
-		y = y + abs(deltaY);
+		x = x + abs((int)deltaX);
+		y = y + abs((int)deltaY);
 	      }
 	    else if ((90 < angle) && (angle < 180))
 	      {
-		x = x + abs(deltaX);
-		y = y - abs(deltaY);
+		x = x + abs((int)deltaX);
+		y = y - abs((int)deltaY);
 	      }
 	    else if ((180 < angle) && (angle < 270))
 	      {
-		x = x - abs(deltaX);
-		y = y - abs(deltaY);
+		x = x - abs((int)deltaX);
+		y = y - abs((int)deltaY);
 	      }
 	    else if ((270 < angle) && (angle < 360))
 	      {
-		x = x - abs(deltaX);
-		y = y + abs(deltaY);
+		x = x - abs((int)deltaX);
+		y = y + abs((int)deltaY);
 	      }
           
 	    MoveToEx(hdc,x,y,NULL);
